Add an interface argument

+++ setupwg.sh
@@ -9,22 +9,27 @@ EXTRA_ALLOWED_IP="0.0.0.0/0"
 INSTALL_SCRIPT_COUNT=1
 
 function runCmd() {
-   echo "executing.. $@" >&2
-   "$@"
+    echo "executing.. $@" >&2
+    "$@"
 }
 
 die() {
-        echo "$PROGRAM: $*" >&2
-        exit 1
+    echo "$PROGRAM: $*" >&2
+    exit 1
+}
+
+function getInterface() {
+    #redirect both stderr and stdout to /dev/null
+    sudo wg show $INTERFACE &> /dev/null
 }
 
 
 function usage() {
     echo "Usage:"
     echo "$0 -> this help"
-    echo "$0 clean -> Cleanup interfaces"
-    echo "$0 create [-s presharedkey]  [-r privatekey] [-l peerpublickey] [-i publicip] [-e edgeip] -> Create interfaces"
-    echo "$0 generate <-i publicip> [-c count] -> Setup local interface and generate 'count' install scripts for peers"
+    echo "$0 clean [ -w wireguardinterface ] -> Cleanup interfaces"
+    echo "$0 create [-s presharedkey] [ -w wireguardinterface ] [ -r privatekey ] [ -l peerpublickey ] [ -i publicip ] [ -e edgeip ] -> Create interfaces"
+    echo "$0 generate < -i publicip > [ -w wireguardinterface ] [ -c count ] -> Setup local interface and generate 'count' install scripts for peers"
     exit 1
 }
 
@@ -33,35 +38,50 @@ if [ "$#" == "0" ]; then
 fi
 
 if [ "$1" == "clean" ]; then
-    echo "Cleaning up.. "
-    rm -f privatekey.script publickey.script presharedkey*.script publickeypeer.script privatekeypeer.script
-    runCmd sudo ip link delete dev $INTERFACE
+    shift 1
+    while getopts ":hw:" o; do
+        case "${o}" in
+            w) INTERFACE=${OPTARG}
+            ;;
+            *|h) usage
+            ;;
+        esac
+    done
+    shift $((OPTIND-1))
+    echo -n "Cleaning up the interface $INTERFACE. Are you sure[y/n]? "
+    read response
+    if [[ "$response" == "y" ]]; then
+        rm -f privatekey.script publickey.script presharedkey*.script publickeypeer.script privatekeypeer.script
+        runCmd sudo ip link delete dev $INTERFACE
+    fi
 else
     if [ "$1" == "generate" ]; then
         generateFlag=1
         shift 1
-        while getopts ":hi:c:" o; do
+        while getopts ":hi:w:c:" o; do
             case "${o}" in
                 i) publicipfromcommandline=${OPTARG}
                 ;;
                 c) INSTALL_SCRIPT_COUNT=${OPTARG}
+                ;;
+                w) INTERFACE=${OPTARG}
                 ;;
                 *|h) usage
                 ;;
             esac
         done
         shift $((OPTIND-1))
+
         if [ -z "$publicipfromcommandline" ]; then
-	    echo "provide the public ip address with -i options"
+            echo "provide the public ip address with -i options"
             usage
         fi
-	if [ "$INSTALL_SCRIPT_COUNT" -lt 0 ]; then 
-           INSTALL_SCRIPT_COUNT=1
-	fi
-	if [ "$INSTALL_SCRIPT_COUNT" -gt 10 ]; then 
-           INSTALL_SCRIPT_COUNT=10
-	fi
-
+        if [ "$INSTALL_SCRIPT_COUNT" -lt 0 ]; then 
+            INSTALL_SCRIPT_COUNT=1
+        fi
+        if [ "$INSTALL_SCRIPT_COUNT" -gt 10 ]; then 
+            INSTALL_SCRIPT_COUNT=10
+        fi
     elif [ "$1" == "create" ]; then
         shift 1
         while getopts ":hs:r:l:i:e:" o; do
@@ -81,6 +101,13 @@ else
             esac
         done
         shift $((OPTIND-1))
+    fi
+
+    getInterface
+    if [[ "$?" == "0" ]]; then
+        echo
+        echo "Interface $INTERFACE exists. This script will stop  !!!"
+        exit 1
     fi
 
     umask 033
@@ -152,41 +179,40 @@ else
             echo -n "Enter the public key of edge:"
             read peerpublickey
         else
-	    peerpublicip=$publicipfromcommandline
+            peerpublicip=$publicipfromcommandline
             publickey=`cat publickey.script`
-	    echo =========================================================
+            echo =========================================================
             echo "public     : $publickey"
             echo "publicipfromcommandline: $publicipfromcommandline"
-	    echo =========================================================
-	    for ii in $(seq 1 $INSTALL_SCRIPT_COUNT)
-	    do
+            echo =========================================================
+            for ii in $(seq 1 $INSTALL_SCRIPT_COUNT); do
                 wg genpsk > presharedkey${ii}.script
                 presharedkey=`cat presharedkey${ii}.script`
-		EDGEIP=${BASEIP}.$((ii+1))
+                EDGEIP=${BASEIP}.$((ii+1))
                 umask 033
                 wg genkey | tee privatekeypeer.script | wg pubkey > publickeypeer.script
                 peerpublickey=`cat publickeypeer.script`
                 peerprivatekey=`cat privatekeypeer.script`
                 echo
-	        echo =========================================================
+                echo =========================================================
                 echo "peerpublic : $peerpublickey"
                 echo "peerprivate: $peerprivatekey"
                 echo "preshared  : $presharedkey"
-	        echo =========================================================
-	        echo "Commands to execute on the peer $ii"
-	        echo =============   Either the following set of commands =======================================================
+                echo =========================================================
+                echo "Commands to execute on the peer $ii"
+                echo =============   Either the following set of commands =======================================================
                 echo "ip link add dev $INTERFACE type wireguard"
                 echo "ip addr add dev $INTERFACE ${EDGEIP}/${MASK}"
                 #echo "ip addr add dev $INTERFACE ${EDGEIP} peer ${MANAGERIP}"
-	        echo "echo '$peerprivatekey' > privatekey.script"
-	        echo "echo '$presharedkey' > presharedkey.script"
-	        echo "sudo wg set $INTERFACE listen-port $LISTENPORT private-key privatekey.script peer $publickey allowed-ips ${MANAGERIP}/$MASK,${EXTRA_ALLOWED_IP} preshared-key presharedkey.script endpoint $peerpublicip:$LISTENPORT persistent-keepalive 20"
+                echo "echo '$peerprivatekey' > privatekey.script"
+                echo "echo '$presharedkey' > presharedkey.script"
+                echo "sudo wg set $INTERFACE listen-port $LISTENPORT private-key privatekey.script peer $publickey allowed-ips ${MANAGERIP}/$MASK,${EXTRA_ALLOWED_IP} preshared-key presharedkey.script endpoint $peerpublicip:$LISTENPORT persistent-keepalive 20"
                 echo "ip link set up dev $INTERFACE"
-		echo ==============  Or if this script is available, then run the script as below\(copy/paste\)=====================
-	        echo "./setupwg.sh create -s $presharedkey -r $peerprivatekey -l $publickey -i $peerpublicip -e $EDGEIP"
-	        echo ==============================================================================================================
+                echo ==============  Or if this script is available, then run the script as below\(copy/paste\)=====================
+                echo "./setupwg.sh create -s $presharedkey -r $peerprivatekey -l $publickey -i $peerpublicip -e $EDGEIP"
+                echo ==============================================================================================================
                 WG_COMMAND="$WG_COMMAND peer $peerpublickey allowed-ips ${EDGEIP}/$MASK,${EXTRA_ALLOWED_IP} preshared-key presharedkey${ii}.script persistent-keepalive 20"
-	    done
+            done
         fi
 
         runCmd $WG_COMMAND
